Initial commit of issue_2191 for OSX
